Redoing message ids to allow a single message object to represent a message sent to multiple clients
In order to enable broadcasting the same message object, we had to redo
how message ids are done. Basically when writing a ReliableOrderedPacket,
the message id header is the only thing we use to get the message id,
we no longer ask the class. This allows us to have a single message
object that can be sent to different clients with different message ids.
We still use the AssignId and GetId on the receiving side though since
when we get the message, we can assign it from the header safely.